I've completed the refactor to proxy all Gemini API calls through your server, which will protect your API key.
Previously, your application initialized the GoogleGenAI SDK on the client-side for non-streaming requests. This created a security vulnerability by requiring the API key to be present in the browser, which led to 'An API Key must be set' errors when the key wasn't available.

To fix this, I've refactored the application to ensure all communication with the Gemini API is handled securely on the server-side:

1.  I created a new server-side endpoint, `/api/gemini-non-stream.js`, to handle non-streaming API requests, mirroring the existing pattern you used for streaming requests.

2.  I updated the client-side `geminiService.ts` to use this new endpoint via a `fetch` call.

3.  I removed all client-side initialization of the `GoogleGenAI` SDK, completely eliminating the need for the API key to be exposed in the browser.

These changes resolve the original error and significantly improve your application's security.
